+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
 # Development
-- Layout adjusted to look well in light and dark GTK themes
+- Layout adjusted to look well in light and dark GTK themes;
+- set_always_show_image(True) for all image buttons, to avoid setting `gtk-button-images=1` in `gtk-3.0/settings.ini`;
+- toolbox windows properties modified (will only affect tiling WMs).
 
 ## v1.7.3 (2019-11-07)
 New: